import_google() no longer uses the deprecated GdataService member auth_token
In gdata < 1.20, GDataService had a member called auth_token which stored the string for the authorization header. In gdata >= 1.20, this was replaced by a token_store, and several Token classes including AuthSubToken as described in http://code.google.com/p/gdata-python-client/wiki/HTTPRequestRefactoring .

The new import_google() will accept either a string token (old-style) or an AuthSubToken (new-style) and deal with it accordingly.